requirements.txt
streamlit==1.25.0  # Streamlit for building the web app
Pillow    # Pillow for image processing